+++ lib/Sema/TypeCheckProtocol.h
@@ -112,8 +112,8 @@ enum class ResolveWitnessResult {
 /// This helper class handles most of the details of checking whether a
 /// given type (\c Adoptee) conforms to a protocol (\c Proto).
 class ConformanceChecker : public WitnessChecker {
-  /// Whether we already suggested adding `@preconcurrency`.
-  bool suggestedPreconcurrency = false;
+  /// Whether we already suggested adding `@preconcurrency` or 'isolated'.
+  bool suggestedPreconcurrencyOrIsolated = false;
 
 public:
   NormalProtocolConformance *Conformance;